+++ Source/Views/PolytempoNetwork/Polytempo_AuxiliaryView.cpp
@@ -90,11 +90,11 @@ void Polytempo_AuxiliaryView::paint (Graphics& g)
         
     }
     
-    g.drawHorizontalLine(320, 0, getWidth());
+    g.drawHorizontalLine(320, 0.0f, (float)getWidth());
     peers.draw(g, Rectangle<int>(10, 330, getWidth() - 20, 100).toFloat());
 
     g.setColour(Colours::grey);
-    g.drawVerticalLine(0, 0, getHeight());
+    g.drawVerticalLine(0, 0.0f, (float)getHeight());
 }
 
 void Polytempo_AuxiliaryView::resized()
@@ -104,7 +104,7 @@ void Polytempo_AuxiliaryView::resized()
     markerTextbox->setBounds(10, yPosition, getWidth() - 20, 66);
     yPosition +=66;
     
-    int buttonWidth = getWidth() * 0.25 - 5;
+    int buttonWidth = (int)(getWidth() * 0.25f - 5.0f);
     int widthCorrection = getWidth() - 20 - buttonWidth * 4;
     imageBackwards->setBounds (10,                   yPosition, buttonWidth, 20);
     markerBackwards->setBounds(10 + buttonWidth * 1, yPosition, buttonWidth, 20);
